글 삭제 로직 수정
