+++ src/components/PageContainer/index.jsx
@@ -1,11 +1,8 @@
 import React, { useEffect, useState } from "react";
 import * as S from "./style";
 import * as C from "../../components";
-import { Link, useNavigate, useParams } from "react-router-dom";
+import { Link, useParams } from "react-router-dom";
 import GetRole from "../../lib/GetRole";
-import { useCallback } from "react";
-import { toast } from "react-toastify";
-import API from "../../apis";
 
 function PageContainer({
   children,
@@ -15,33 +12,13 @@ function PageContainer({
   hasPostButton,
   hasHistoryButton,
   hasDeleteButton,
-  url
+  url,
+  onClick
 }) {
   const { id } = useParams();
   const data = GetRole();
-  const navigate = useNavigate();
 
   const [roleUrl, setRoleUrl] = useState("");
-
-  const boardDelete = useCallback(async () => {
-    try {
-      await API.delete(`/${roleUrl}/board/${id}`);
-
-      toast.success("삭제 성공");
-      navigate("/");
-    } catch (e) {
-      console.log(e);
-      toast.error("글 삭제에 실패했습니다.");
-    }
-  }, [roleUrl, id, navigate]);
-
-  const handleDelete = () => {
-    const shouldDelete = window.confirm("정말로 삭제하시겠습니까?");
-
-    if (shouldDelete) {
-      boardDelete();
-    }
-  };
 
   useEffect(() => {
     if (data === "관리자") {
@@ -73,15 +50,7 @@ function PageContainer({
                     <C.ContentsButton>역사</C.ContentsButton>
                   </Link>
                 )}
-                {hasDeleteButton && (
-                  <button
-                    onClick={() => {
-                      handleDelete();
-                    }}
-                  >
-                    삭제
-                  </button>
-                )}
+                {hasDeleteButton && <button onClick={onClick}>삭제</button>}
               </>
             </S.ContentsButtonContainer>
           </C.WhiteContainer>
