+++ src/pages/BoardDetail/index.jsx
@@ -1,8 +1,11 @@
 import React, { useEffect, useState } from "react";
 import * as C from "../../components";
 import { useFetch } from "../../Hooks";
-import { useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import GetRole from "../../lib/GetRole";
+import { useCallback } from "react";
+import API from "../../apis";
+import { toast } from "react-toastify";
 
 const BoardDetail = () => {
   const data = GetRole();
@@ -44,6 +47,28 @@ const BoardDetail = () => {
     }
   }, [roleUrl]);
 
+  const navigate = useNavigate();
+
+  const boardDelete = useCallback(async () => {
+    try {
+      await API.delete(`/${roleUrl}/board/${id}`);
+      toast.success("삭제 성공");
+
+      navigate("/");
+    } catch (e) {
+      console.log(e);
+      toast.error("글 삭제에 실패했습니다.");
+    }
+  }, [roleUrl, navigate, id]);
+
+  const handleDelete = () => {
+    const shouldDelete = window.confirm("정말로 삭제하시겠습니까?");
+
+    if (shouldDelete) {
+      boardDelete();
+    }
+  };
+
   return (
     <>
       <C.RecentModified />
@@ -55,6 +80,7 @@ const BoardDetail = () => {
         {...(data === "관리자"
           ? { hasDeleteButton: true }
           : { hasDeleteButton: false })}
+        onClick={handleDelete}
         sort="글"
       >
         <C.Explanation>
